backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langgraph
langchain
langchain-openai
langchain-tavily
tavily-python
deepagents
pydantic
sse-starlette